Performance improvements Internal refactoring
