+++ lib/src/bulk_download/tile_loops/shared.dart
@@ -1,0 +1,27 @@
+// Copyright © Luka S (JaffaKetchup) under GPL-v3
+// A full license can be found at .\LICENSE
+
+import 'dart:math';
+
+import 'package:flutter_map/flutter_map.dart';
+import 'package:flutter_map/flutter_map.dart' hide Polygon;
+import 'package:latlong2/latlong.dart';
+
+part 'count.dart';
+part 'generate.dart';
+
+class _Polygon {
+  final CustomPoint<num> nw;
+  final CustomPoint<num> ne;
+  final CustomPoint<num> se;
+  final CustomPoint<num> sw;
+
+  _Polygon(this.nw, this.ne, this.se, this.sw);
+
+  List<CustomPoint<num>> get points => [nw, ne, se, sw];
+}
+
+extension on List<double> {
+  double get minNum => reduce(min);
+  double get maxNum => reduce(max);
+}
