+++ lib/src/errors/browsing.dart
@@ -1,3 +1,6 @@
+// Copyright © Luka S (JaffaKetchup) under GPL-v3
+// A full license can be found at .\LICENSE
+
 import 'package:meta/meta.dart';
 
 import '../../flutter_map_tile_caching.dart';
