+++ lib/src/errors/damaged_store.dart
@@ -1,3 +1,6 @@
+// Copyright © Luka S (JaffaKetchup) under GPL-v3
+// A full license can be found at .\LICENSE
+
 import 'package:meta/meta.dart';
 
 /// An [Exception] indicating that an operation was attempted on a damaged store
