+++ lib/src/errors/initialisation.dart
@@ -1,3 +1,6 @@
+// Copyright © Luka S (JaffaKetchup) under GPL-v3
+// A full license can be found at .\LICENSE
+
 import 'package:meta/meta.dart';
 
 /// An [Exception] raised when FMTC failed to initialise
