+++ lib/src/errors/store_not_ready.dart
@@ -1,3 +1,6 @@
+// Copyright © Luka S (JaffaKetchup) under GPL-v3
+// A full license can be found at .\LICENSE
+
 import 'package:meta/meta.dart';
 
 /// An [Error] indicating that a store did not exist when it was expected to
