+++ lib/src/misc/store_db_impl.dart
@@ -1,3 +1,6 @@
+// Copyright © Luka S (JaffaKetchup) under GPL-v3
+// A full license can be found at .\LICENSE
+
 part of flutter_map_tile_caching;
 
 abstract class _StoreDb {
